refactor - add board class
